requirements.txt
faiss-cpu
Deprecated==1.2.14

#代码解析相关
tree-sitter==0.24.0 # Java/Kotlin 代码解析
tree-sitter-java==0.23.5 # Java 代码解析
tree-sitter-kotlin==1.1.0
networkx==3.3 # 图处理

#向量嵌入相关
sentence-transformers==2.7.0 # 文本嵌入
torch==2.7.1+cpu # 向量嵌入
torchaudio==2.7.1+cpu # 向量嵌入
torchvision==0.22.1+cpu # 向量嵌入

sqlalchemy==2.0.25 # sqllite ORM

# 大模型交互
openai==1.3.6
tiktoken==0.5.2 # 计算token

# 工具类库
loguru==0.7.0 # 日志
pydantic-settings==2.0.2 # 配置文件